string are now hashed instead of assigning id's to each string. This should solve the issue with the linking of multiple libraries with the same identifers.

+++ runtime/include/kitpl/core.hpp
@@ -37,7 +37,7 @@ namespace kit
 	class Object;
 
 	// message id
-	typedef unsigned int MsgId;
+	typedef unsigned long MsgId;
 
 	// function pointer prototype
 //	typedef Object* (*FuncPtr)(Object*, MsgId, ... );
